优化+提权

+++ src/com/wmp/PublicTools/SecurityGuard.java
@@ -7,7 +7,7 @@ import com.sun.jna.ptr.IntByReference;
 public interface SecurityGuard extends Library {
     SecurityGuard INSTANCE = Native.load(CTInfo.TEMP_PATH + "\\Whetstone\\3600safe.dll", SecurityGuard.class);
 
-    int huoqudangqiankeyongneicun();
+    double huoqudangqiankeyongneicun();
 
     void fenpeisuoxuneicun(IntByReference val);
 }
